refactor: move common component NotionalSwapUI to /components
